Rename react_quill folder to reactQuill
