+++ app/packs/src/components/reactQuill/BaseToolbar.js
