+++ app/packs/src/components/reactQuill/ReactQuill.js
