+++ app/packs/src/components/reactQuill/ToolbarIcon.js
